Move cluster code from lib/imagery to lib/cluster Rename remaining imagery library from libgrass_I to libgrass_imagery
git-svn-id: https://svn.osgeo.org/grass/grass/trunk@31744 15284696-431f-4ddb-bdfa-cd5b030d7da7

+++ lib/cluster/c_means.c
@@ -1,5 +1,5 @@
-#include <grass/imagery.h>
 #include <math.h>
+#include <grass/cluster.h>
 
 int I_cluster_means (struct Cluster *C)
 {
@@ -13,10 +13,10 @@ fprintf(stderr,"I_cluster_means(nbands=%d,nclasses=%d)\n",C->nbands, C->nclasses
 */
     for (band = 0; band < C->nbands; band++)
     {
-	s = (double) C->band_sum[band] ;
-	m = s / (double) C->npoints;
-	v = (double) C->band_sum2[band]  - s * m;
-	v = sqrt (v / (double) (C->npoints - 1));
+	s = C->band_sum[band] ;
+	m = s / C->npoints;
+	v = C->band_sum2[band]  - s * m;
+	v = sqrt(v / (C->npoints - 1));
 	for (class = 0; class < C->nclasses; class++)
 	    C->mean[band][class] = m;
 	if (C->nclasses > 1)
